requirements.txt
edge-tts>=6.1.0
ffmpeg-python>=0.2.0
asyncio-throttle>=1.0.2
numpy>=1.24.0